Added a way to disable the clr object (#116)
* Added a way to dissable the clr object

Summary of changes:
For security reasons, I need a way to disable the
CLR functionality. To achieve this goal, I added a new parameter to the
`LuaCompileOptions` class called `ClrEnabled`. To preserve existing
functionality, `ClrEnabled` is set to true by default. I've also added
ways to test this new functionality through the unit tests and via the
`NeoCmd` utility: I've also added a  new parameter at the bottom of the
CLR documentation file to explain how disabling CLR works.

List of changes:
* Added a new parameter called `ClrEnabled` to the `LuaCompileOptions`
class (set to true by default).
* Added a check for the `ClrEnabled` option in the `ParsePrefix`
function of the `Parser` class.
* Added `TestClrEnabled` and
`TestClrDisabled` in the `Parser.cs` unit test file.
* Added the
`:clron` and `clroff` to the `NeoCmd` utility.
* Added documentation for the new `NeoCmd` commands;
* Added documentation for the new `ClrEnabled` parameter in the `CLR`
documentation file.

* why did this not get checked in...

Somehow this change wasn't check in

* Fixes naming convention in NeoCmd

Refactores the name of the `ClrEnabled` static variable in the
program.cs of the NeoCmd `clrEnabled` as requested in
https://github.com/neolithos/neolua/pull/116#discussion_r453060943

* removed broken from documentation

Removed a line out of the `Disabling clr access` documentation section.
This permits the code to be dropped in for testing purposes.